Initial commit
Floating point numbers switched from 'double' to 'float128' (Boost
Multiprecision library used). Known issues — computations slow down by
at least 2 times. Calculation accuracy is not checked yet as well.

+++ src/Widgets/elementstablewidget.h
@@ -25,13 +25,13 @@ class ElementsTableWidget : public QWidget
     public:
         explicit ElementsTableWidget(ACL::Data::ElementsContentUnits contentsUnits,
                                      QWidget *parent = 0);
-        ElementsTableWidget(std::vector<double> contents,
+        ElementsTableWidget(std::vector<float128> contents,
                             ACL::Data::ElementsContentUnits contentsUnits,
-                            std::vector<double> weights,
+                            std::vector<float128> weights,
                             QWidget *parent = 0);
         ~ElementsTableWidget();
-        std::vector<double> getElementsWeights();
-        std::vector<double> getElementsContent();
+        std::vector<float128> getElementsWeights();
+        std::vector<float128> getElementsContent();
         ACL::Data::ElementsContentUnits getElementsContentUnits();
 
     private:
